requirements.txt
# Core Web Framework
fastapi
uvicorn

# Environment & Config
python-dotenv
pydantic
pydantic-settings

# Database & ORM
supabase

# AI/LLM - Gemini (sesuai dengan API key di .env)
langchain
langchain-google-genai

# Document Processing

# Utilities

# Testing & Development
pytest
pytest-asyncio

# Logging & Monitoring
python-json-logger